+++ test/ttmlir/Dialect/TTNN/trace/creation_no_consteval.mlir
@@ -21,7 +21,7 @@ module {
     // CHECK-NOT: "ttnn.add"
     // CHECK-NEXT: %[[TRACE_RESULT:.+]] = "ttnn.capture_or_execute_trace"(%[[GET_DEVICE]], %1, %2) <{capture_callee = @run_and_capture_trace_0_creation_ops, execute_callee = @execute_trace_0_creation_ops}>
     // CHECK: return %[[TRACE_RESULT]]
-    %0 = "ttir.zeros"() <{shape = array<i32: 4, 4>, dtype = i64}> : () -> tensor<4x4xbf16>
+    %0 = "ttir.zeros"() <{shape = array<i32: 4, 4>, dtype = bf16}> : () -> tensor<4x4xbf16>
     %1 = "ttir.ones"() <{shape = array<i32: 4, 4>, dtype = bf16}> : () -> tensor<4x4xbf16>
 
     %2 = ttir.empty() : tensor<4x4xbf16>
